fix(quarkus): cannot include camel-quarkus-cxf-soap
Fix #3703
